# backend/internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: SpendSense API
  version: "1.0.0"
  description: API for SpendSense Phase 1 (expenses, incomes, wallets, categories, auth)
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/expenses:
    post:
      summary: Create expense
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Created

    get:
      summary: List expenses (pagination)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: pagination
          schema:
            type: string
      responses:
        '200':
          description: List
  /api/v1/expenses/{id}:
    get:
      summary: Get expense by id
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: expense

    put:
      summary: Update expense
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExpenseRequest'
      responses:
        '200':
          description: Updated

    delete:
      summary: Soft-delete expense
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
...
  /api/v1/wallets:
    post:
      summary: Create wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalletRequest'
      responses:
        '201':
          description: Created

    get:
      summary: List wallets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /api/v1/wallets/transfer:
    post:
      summary: Transfer between wallets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalletTransferRequest'
      responses:
        '201':
          description: Created
  /api/v1/incomes:
    post:
      summary: Create income
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncomeRequest'
      responses:
        '201':
          description: Created

    get:
      summary: List incomes
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: pagination
          schema:
            type: string
      responses:
        '200':
          description: OK
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateExpenseRequest:
      type: object
      required: [wallet_id, amount, currency, date]
      properties:
        wallet_id:
          type: integer
        amount:
          type: number
          format: double
          description: Decimal amount with 2-digit precision.
        currency:
          type: string
        category_id:
          type: integer
        date:
          type: string
          format: date
        merchant:
          type: string
        notes:
          type: string
    UpdateExpenseRequest:
      type: object
      properties:
        merchant:
          type: string
        notes:
          type: string
...
    CreateWalletRequest:
      type: object
      required: [name, wallet_type, opening_balance, currency]
      properties:
        name:
          type: string
        wallet_type:
          type: string
          enum: [CASH, MOBILE_WALLET, BANK, CARD]
        provider:
          type: string
          nullable: true
        opening_balance:
          type: number
          format: double
        currency:
          type: string
    CreateIncomeRequest:
      type: object
      required: [wallet_id, source_name, amount, currency, income_date]
      properties:
        wallet_id:
          type: integer
        category_id:
          type: integer
          nullable: true
        source_name:
          type: string
        amount:
          type: number
          format: double
          description: Decimal amount with 2-digit precision.
        currency:
          type: string
        income_date:
          type: string
          format: date
        notes:
          type: string
    CreateWalletTransferRequest:
      type: object
      required: [from_wallet_id, to_wallet_id, amount, currency, transfer_date]
      properties:
        from_wallet_id:
          type: string
        to_wallet_id:
          type: string
        amount:
          type: number
          format: double
          description: Source wallet amount with 2-digit precision.
        exchange_rate:
          type: number
          format: double
          description: Required when source and destination wallets use different currencies.
        fee_amount:
          type: number
          format: double
        currency:
          type: string
        transfer_date:
          type: string
          format: date
        notes:
          type: string
...